Cleaned up indexing and dataset generation workflow.

+++ src/main/java/nl/andrewl/email_indexer/gen/EmailIndexGenerator.java
@@ -1,12 +1,10 @@
 package nl.andrewl.email_indexer.gen;
 
 import nl.andrewl.email_indexer.data.EmailDataset;
+import nl.andrewl.email_indexer.data.EmailEntryPreview;
 import nl.andrewl.email_indexer.data.EmailRepository;
 import nl.andrewl.email_indexer.data.EmailSearchResult;
 import nl.andrewl.email_indexer.data.util.FileUtils;
-import nl.andrewl.mboxparser.CompositeEmailHandler;
-import nl.andrewl.mboxparser.EmailHandler;
-import nl.andrewl.mboxparser.MBoxParser;
 import org.apache.lucene.analysis.Analyzer;
 import org.apache.lucene.analysis.standard.StandardAnalyzer;
 import org.apache.lucene.document.Document;
@@ -20,9 +18,6 @@ import org.apache.lucene.store.FSDirectory;
 
 import java.io.IOException;
 import java.nio.file.Files;
-import java.nio.file.LinkOption;
-import java.nio.file.Path;
-import java.util.Collection;
 import java.util.HashSet;
 import java.util.Set;
 import java.util.function.Consumer;
@@ -32,46 +27,18 @@ import java.util.function.Consumer;
  */
 public class EmailIndexGenerator {
 	/**
-	 * Generates an email index and provides all indexed emails to any supplied
-	 * email handlers for processing.
-	 * @param inputDirs Directories to parse mbox files from.
-	 * @param outputDir The directory to place the index data.
-	 * @param messageConsumer A consumer to accept messages emitted during the
-	 *                        generation.
-	 * @param emailHandlers Handlers that will process any indexed emails.
-	 * @throws IOException If an error occurs while reading or writing.
-	 */
-	public void generateIndex(Collection<Path> inputDirs, Path outputDir, Consumer<String> messageConsumer, EmailHandler... emailHandlers) throws IOException {
-		Files.createDirectories(outputDir);
-		Analyzer analyzer = new StandardAnalyzer();
-		IndexWriterConfig config = new IndexWriterConfig(analyzer);
-		try (
-				Directory emailDirectory = FSDirectory.open(outputDir);
-				IndexWriter emailIndexWriter = new IndexWriter(emailDirectory, config)
-		) {
-			// Create a composite handler that has a collection of handlers that are called for each email parsed.
-			var compositeHandler = new CompositeEmailHandler(new IndexingEmailHandler(emailIndexWriter));
-			for (var h : emailHandlers) compositeHandler.withHandler(h);
-			// Wrap the composite handler in a sanitizing handler to filter out plain junk.
-			MBoxParser parser = new MBoxParser(new SanitizingEmailHandler(compositeHandler));
-			for (var dir : inputDirs) {
-				parseRecursive(dir, parser, messageConsumer);
-			}
-			messageConsumer.accept("Indexing complete.");
-		}
-	}
-
-	/**
-	 * Removes and regenerates the indexes for a dataset, based entirely on
-	 * non-hidden emails.
+	 * Generates the indexes for a dataset, based entirely on non-hidden emails.
 	 * @param dataset The dataset to index.
 	 * @param messageConsumer A consumer to accept messages emitted during the
 	 *                        generation.
 	 * @throws IOException If an error occurs while reading or writing.
 	 */
-	public void regenerateIndex(EmailDataset dataset, Consumer<String> messageConsumer) throws IOException {
-		FileUtils.deleteFiles(dataset.getIndexDir());
-		messageConsumer.accept("Removed existing index files.");
+	public void generateIndex(EmailDataset dataset, Consumer<String> messageConsumer) throws IOException {
+		if (Files.exists(dataset.getIndexDir())) {
+			FileUtils.deleteFiles(dataset.getIndexDir());
+		} else {
+			Files.createDirectory(dataset.getIndexDir());
+		}
 		Analyzer analyzer = new StandardAnalyzer();
 		IndexWriterConfig config = new IndexWriterConfig(analyzer);
 		EmailRepository repo = new EmailRepository(dataset);
@@ -88,17 +55,7 @@ public class EmailIndexGenerator {
 				for (var email : result.emails()) {
 					if (!emailIds.contains(email.messageId())) {
 						emailIds.add(email.messageId());
-						repo.getBody(email.messageId()).ifPresent(body -> {
-							Document doc = new Document();
-							doc.add(new StringField("id", email.messageId(), Field.Store.YES));
-							doc.add(new StringField("subject", email.subject(), Field.Store.NO));
-							doc.add(new Field("body", body, TextField.TYPE_NOT_STORED));
-							try {
-								emailIndexWriter.addDocument(doc);
-							} catch (IOException e) {
-								e.printStackTrace();
-							}
-						});
+						addToIndex(email, repo, emailIndexWriter);
 					}
 				}
 				if (result.hasNextPage()) {
@@ -112,17 +69,20 @@ public class EmailIndexGenerator {
 		}
 	}
 
-	private void parseRecursive(Path dir, MBoxParser parser, Consumer<String> messageConsumer) throws IOException {
-		messageConsumer.accept("Parsing directory: " + dir);
-		try (var s = Files.list(dir)) {
-			for (var p : s.toList()) {
-				if (Files.isDirectory(p, LinkOption.NOFOLLOW_LINKS)) {
-					parseRecursive(p, parser, messageConsumer);
-				} else if (Files.isRegularFile(p, LinkOption.NOFOLLOW_LINKS) && Files.isReadable(p)) {
-					messageConsumer.accept("Parsing file: " + p);
-					parser.parse(p);
-				}
+	private void addToIndex(EmailEntryPreview email, EmailRepository repo, IndexWriter emailIndexWriter) {
+		repo.getBody(email.messageId()).ifPresent(body -> {
+			Document doc = new Document();
+			doc.add(new StringField("id", email.messageId(), Field.Store.YES));
+			doc.add(new StringField("subject", email.subject(), Field.Store.NO));
+			doc.add(new Field("body", body, TextField.TYPE_NOT_STORED));
+			repo.findRootEmailByChildId(email.messageId()).ifPresent(rootEmail -> {
+				doc.add(new StringField("rootId", rootEmail.messageId(), Field.Store.YES));
+			});
+			try {
+				emailIndexWriter.addDocument(doc);
+			} catch (IOException e) {
+				e.printStackTrace();
 			}
-		}
+		});
 	}
 }
